requirements.txt
-i https://pypi.org/simple
altair==5.5.0; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
attrs==25.3.0; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.1.31; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
docstring-parser==0.16; python_version >= '3.6' and python_version < '4.0'
gitdb==4.0.12; python_version >= '3.7'
gitpython==3.1.44; python_version >= '3.7'
google-api-core[grpc]==2.24.2; python_version >= '3.7'
google-auth==2.38.0; python_version >= '3.7'
google-cloud-aiplatform[all]; python_version >= '3.8'
google-cloud-bigquery==3.31.0; python_version >= '3.9'
google-cloud-core==2.4.3; python_version >= '3.7'
google-cloud-discoveryengine==0.13.8; python_version >= '3.7'
google-cloud-resource-manager==1.14.2; python_version >= '3.7'
google-cloud-storage==2.19.0; python_version >= '3.7'
google-cloud-vision==3.10.1; python_version >= '3.7'
google-crc32c==1.7.1; python_version >= '3.9'
google-genai==1.9.0; python_version >= '3.9'
google-resumable-media==2.7.2; python_version >= '3.7'
googleapis-common-protos[grpc]==1.69.2; python_version >= '3.7'
grpc-google-iam-v1==0.14.2; python_version >= '3.7'
grpcio==1.71.0; python_version >= '3.9'
grpcio-status==1.71.0; python_version >= '3.9'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
httpx-sse==0.4.0; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
jinja2==3.1.6; python_version >= '3.7'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2024.10.1; python_version >= '3.9'
langchain==0.3.23; python_version >= '3.9' and python_version < '4.0'
langchain-core==0.3.51; python_version >= '3.9' and python_version < '4.0'
langchain-google-vertexai==2.0.19; python_version >= '3.9' and python_version < '4.0'
langchain-text-splitters==0.3.8; python_version >= '3.9' and python_version < '4.0'
langgraph==0.3.25; python_full_version >= '3.9.0' and python_version < '4.0'
langgraph-checkpoint==2.0.24; python_full_version >= '3.9.0' and python_full_version < '4.0.0'
langgraph-prebuilt==0.1.8; python_full_version >= '3.9.0' and python_full_version < '4.0.0'
langgraph-sdk==0.1.61; python_full_version >= '3.9.0' and python_full_version < '4.0.0'
langsmith==0.3.28; python_version >= '3.9' and python_version < '4.0'
markupsafe==3.0.2; python_version >= '3.9'
narwhals==1.34.1; python_version >= '3.8'
numpy==2.2.4; python_version >= '3.10'
orjson==3.10.16; python_version >= '3.9'
ormsgpack==1.9.1; python_version >= '3.9'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
pdf2image==1.17.0
pillow==11.1.0; python_version >= '3.9'
proto-plus==1.26.1; python_version >= '3.7'
protobuf==5.29.4; python_version >= '3.8'
pyarrow==19.0.1; python_version >= '3.9'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pydantic==2.9.2; python_version >= '3.8'
pydantic-core==2.23.4; python_version >= '3.8'
pydeck==0.9.1; python_version >= '3.8'
pypdf2==3.0.1; python_version >= '3.6'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2025.2
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.36.2; python_version >= '3.9'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rpds-py==0.24.0; python_version >= '3.9'
rsa==4.9; python_version >= '3.6' and python_version < '4'
shapely==2.1.0; python_version >= '3.10'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
smmap==5.0.2; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.40; python_version >= '3.7'
streamlit==1.44.1; python_version >= '3.9' and python_full_version != '3.9.7'
tavily-python==0.5.4; python_version >= '3.6'
tenacity==9.1.2; python_version >= '3.9'
tiktoken==0.9.0; python_version >= '3.9'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
tornado==6.4.2; python_version >= '3.8'
typing-extensions==4.13.1; python_version >= '3.8'
typing-inspection==0.4.0; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
urllib3==2.3.0; python_version >= '3.9'
validators==0.34.0; python_version >= '3.8'
websockets==15.0.1; python_version >= '3.9'
xxhash==3.5.0; python_version >= '3.7'
zstandard==0.23.0; python_version >= '3.8'
